isochrones/requirements.txt
ai.cs==1.0.7
aiofiles==0.5.0
argon2-cffi==20.1.0
astroid==2.4.2
astropy==4.0.1.post1
attrs==20.2.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
blinker==1.4
boto3==1.14.62
botocore==1.17.62
bresenham==0.2.1
certifi==2020.6.20
cffi==1.14.2
cftime==1.2.1
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
coloredlogs==14.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
Fiona==1.8.17
Flask==1.1.2
geopandas==0.8.1
h11==0.10.0
h2==3.2.0
hpack==3.0.0
humanfriendly==8.2
Hypercorn==0.10.2
hyperframe==5.2.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
-e .
isort==5.4.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
netCDF4==1.5.4
notebook==6.1.3
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
priority==1.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.6.0
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.16.0
pyshp==2.1.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
s3transfer==0.3.3
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==0.15.0
sqlalchemy==1.4.32